tock-register-interface: document & move interface traits
Moves the Readable, Writeable and ReadeWriteable traits to a separate
module `interfaces`. Further adds a documentation comment to this
module, outlining the reasoning behind this interface, how it can be
used to create custom register types, and how deriving ReadWriteable
works.

The introduced documentation comment's code examples also serve as
documentation tests, which will test implementing custom types using
these traits (which implicitly tests interface compatibility), as well
as the ReadWriteable implementation logic on ReadWrite and Aliased
types using an asserted compilation failure.
