+++ chips/litex/src/litex_registers.rs
@@ -31,11 +31,11 @@
 //! }`, `BaseWidth` in `{ 32 }`.
 
 use core::marker::PhantomData;
+use tock_registers::interfaces::{Readable, Writeable};
 pub use tock_registers::register_bitfields;
 use tock_registers::registers::{
     Field, FieldValue, IntLike as TRIntLike, LocalRegisterCopy, ReadOnly as TRReadOnly,
-    ReadWrite as TRReadWrite, Readable, RegisterLongName, TryFromValue, WriteOnly as TRWriteOnly,
-    Writeable,
+    ReadWrite as TRReadWrite, RegisterLongName, TryFromValue, WriteOnly as TRWriteOnly,
 };
 
 /// Extension of the `tock_registers` `IntLike` trait.
