+++ libraries/riscv-csr/src/csr.rs
@@ -2,7 +2,8 @@
 
 use core::marker::PhantomData;
 
-use tock_registers::registers::{Field, IntLike, Readable, RegisterLongName, Writeable};
+use tock_registers::interfaces::{Readable, Writeable};
+use tock_registers::registers::{Field, IntLike, RegisterLongName};
 
 pub const MINSTRETH: usize = 0xB82;
 pub const MINSTRET: usize = 0xB02;
