+++ libraries/tock-register-interface/src/lib.rs
@@ -5,5 +5,6 @@
 #![feature(const_fn_trait_bound)]
 #![no_std]
 
+pub mod interfaces;
 pub mod macros;
 pub mod registers;
